Restructured Files, fixed router issue
