+++ frontend/src/components/TopNav/TopNav.js
@@ -1,15 +1,17 @@
 import React from 'react';
 import './TopNav.css';
-import { Avatar, IconButton, Link } from '@mui/material';
+import { Link } from 'react-router-dom';
+import { Avatar, IconButton } from '@mui/material';
 import InboxIcon from '@mui/icons-material/Inbox';
 import SettingsIcon from '@mui/icons-material/Settings';
 import SearchIcon from '@mui/icons-material/Search';
+
 
 function TopNav(props){
     return(
     <div className='TopNav'>
         <div className='TopNav_logo'>
-            <Link to='/'>
+            <Link to='/login'>
                 <img src = 'logo-stackoverflow.png' className='TopNav_logo_img'/>
             </Link>
         </div>
@@ -23,7 +25,7 @@ function TopNav(props){
         <div className='TopNav_options'>
 
             <IconButton>
-                <Link to='/user-profile'>
+                <Link to="/UserProfile">
                     <Avatar src = {props.avatar_src} className='TopNav_svg'/>
                 </Link>
             </IconButton>
