--- frontend/src/routes/Questions.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Questions() {
-    return (
-        <div>
-            <h2> Questions </h2>
-        </div>
-    )
-}
-
-export default Questions
