+++ frontend/src/routes/UserProfile/Userprofile.js
@@ -56,4 +56,4 @@ function UserProfile(props) {
   )
 }
 
-export default UserProfile
+export default UserProfile;
