--- frontend/src/routes/Users.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Users() {
-    return (
-        <div>
-            <h2> Username, Avatar xd</h2>
-        </div>
-    )
-}
-
-export default Users
